% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_tint}
\alias{theme_tint}
\title{A 'Tufte'-inspired 'ggplot2' theme}
\usage{
theme_tint(base_size = 11, ticks = TRUE)
}
\arguments{
\item{base_size}{An integer value for the base tick size, default
is 11.}

\item{ticks}{A logical value indicating if ticks should be drawn,
default is \code{TRUE}.}
}
\value{
A modified theme for ggplot2 use
}
\description{
A 'Tufte'-inspired 'ggplot2' theme for the 'tint' package
}
\details{
This theme is borrows heavily from a similar theme in \pkg{ggtufte}
by Jeffrey Arnold (but removes his font processing as it conflicts
with ours).
}
\author{
Dirk Eddelbuettel modifying earlier work by Jeffrey Arnold
}
