% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tint.R, R/utils.R
\name{tint}
\alias{margin_note}
\alias{newthought}
\alias{quote_footer}
\alias{sans_serif}
\alias{tint}
\title{Tint Is Not Tufte -- A Tufte-Inspired HTML Format}
\usage{
tint(...)

newthought(text)

margin_note(text, icon = "&#8853;")

quote_footer(text)

sans_serif(text)
}
\arguments{
\item{...}{Other arguments to be passed to
\code{\link{tint}} (note you cannot use the \code{theme} argument in
\code{tint}; this arguments has been set internally)}

\item{text}{A character string to be presented as a \dQuote{new thought}
(using small caps), or a margin note, or a footer of a quote}

\item{icon}{A character string to indicate there is a hidden margin note when
the page width is too narrow (by default it is a circled plus sign)}
}
\description{
A templates for creating handouts according to the style of Edward R. Tufte and
Richard Feynman, but with an updated font choice.
}
\details{
\code{tint} provides the HTML format based on the Tufte CSS
  \url{https://edwardtufte.github.io/tufte-css/} with fonts set according to
\url{http://nogginfuel.com/envisioned-css/}

\code{newthought()} can be used in inline R expressions in R
  Markdown (e.g. \samp{`r newthought(Some text)`}), and it works for both
  HTML (\samp{<span class="newthought">text</span>}) and PDF
  (\samp{\\newthought{text}}) output.

\code{margin_note()} can be used in inline R expressions to write a
  margin note (like a sidenote but not numbered).

\code{quote_footer()} formats text as the footer of a quote. It puts
  \code{text} in \samp{<footer></footer>} for HTML output, and
  after \samp{\\hfill} for LaTeX output (to right-align text).

\code{sans_serif()} applies sans-serif fonts to \code{text}.
}
\examples{
\dontrun{library(tint)}

newthought('In this section')
}
\references{
See \url{http://rstudio.github.io/tufte} for an example.
}

